moves user header to own component and puts it in Profile
