+++ src/routes/UserProfile.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import Container from 'components/Container';
 import Button from 'components/Button';
+import ProfileHeader from 'components/ProfileHeader';
 import { logout } from 'utils/AuthService';
 
 const UserProfile = ({ userResult: { loading, user }, ...props }) => {
@@ -14,17 +15,18 @@ const UserProfile = ({ userResult: { loading, user }, ...props }) => {
       </Container>
     );
   }
-  const { fullName, userName, email, createdAt } = user;
+  const { userName, email, createdAt } = user;
   const joinedDate = new Date(createdAt).getFullYear();
   return (
-    <Container>
-      <h1>Profile</h1>
-      <p>Username: @{userName}</p>
-      <p>Full Name: {fullName}</p>
-      <p>Email: {email}</p>
-      <p>Year Joined: {joinedDate}</p>
-      <Button onClick={logout}>Logout</Button>
-    </Container>
+    <div>
+      <ProfileHeader user={user} />
+      <Container>
+        <p>Username: @{userName}</p>
+        <p>Email: {email}</p>
+        <p>Year Joined: {joinedDate}</p>
+        <Button onClick={logout}>Logout</Button>
+      </Container>
+    </div>
   );
 };
